Migrate native addon to Node-API 8

+++ packages/native-with-source/src/deps/PDFWriter/InputCharStringDecodeStream.cpp
@@ -39,10 +39,10 @@ void InputCharStringDecodeStream::Assign(IByteReader* inReadFrom,unsigned long i
 }
 
 
-static const int CONSTANT_1 = 52845;
-static const int CONSTANT_2 = 22719;
-static const int RANDOMIZER_INIT = 4330;
-static const int RANDOMIZER_MODULU_VAL = 65536;
+static const unsigned int CONSTANT_1 = 52845;
+static const unsigned int CONSTANT_2 = 22719;
+static const unsigned int RANDOMIZER_INIT = 4330;
+static const unsigned int RANDOMIZER_MODULU_VAL = 65536;
 
 void InputCharStringDecodeStream::InitializeCharStringDecode(unsigned long inLenIV)
 {
@@ -73,7 +73,7 @@ EStatusCode InputCharStringDecodeStream::ReadDecodedByte(Byte& outByte)
 Byte InputCharStringDecodeStream::DecodeByte(Byte inByteToDecode)
 {
 	Byte result = (Byte)(inByteToDecode ^ (mRandomizer >> 8));
-	mRandomizer = (unsigned short)(((unsigned int)(inByteToDecode + mRandomizer) * CONSTANT_1 + CONSTANT_2) % RANDOMIZER_MODULU_VAL);
+	mRandomizer = (unsigned short)(((inByteToDecode + mRandomizer)* CONSTANT_1 + CONSTANT_2) % RANDOMIZER_MODULU_VAL);
 	return result;
 }
 
@@ -97,4 +97,5 @@ bool InputCharStringDecodeStream::NotEnded()
 {
 	return mReadFrom->NotEnded();
 }
+
